Downgrade iframe scroll
In iframe hash scroll does not work for hashes containing spaces because
hashes are not decoded. Moreover, the load of the page with anchor does
not work if the page renders the anchors in JQuery ready() method. The
reason is that the scroll is done on the load event but the ready is
triggered later. However, in the version 3.2.2 while it contained other
problems in this area, it works with these pages implemented with
JavaScript even if they trigger render in JQuery ready().